+++ packages/geometry-utils/src/worker-flow/index.ts
@@ -20,7 +20,7 @@ export default function calculate(config: WorkerConfig, buffer: ArrayBuffer): Ar
 
     const polygonCount: number = config.polygons.length;
     const view: DataView = new DataView(buffer);
-    const dataType: THREAD_TYPE = view.getFloat64(0) as THREAD_TYPE;
+    const dataType: THREAD_TYPE = view.getUint32(0) as THREAD_TYPE;
     const isPair: boolean = dataType === THREAD_TYPE.PAIR;
     const result: Float64Array = isPair ? pairData(buffer, config) : placePaths(buffer, config);
